+++ src/main/java/org/example/hana/recruitment/repository/ApplicationRepository.java
@@ -1,4 +1,4 @@
-package org.example.hana.recruitment.db;
+package org.example.hana.recruitment.repository;
 
 import org.example.hana.recruitment.entity.Application;
 import org.springframework.data.jpa.repository.JpaRepository;
