+++ src/main/java/org/example/hana/recruitment/repository/RecruitmentPostRepository.java
@@ -1,4 +1,4 @@
-package org.example.hana.recruitment.db;
+package org.example.hana.recruitment.repository;
 
 import org.example.hana.recruitment.entity.RecruitmentPost;
 import org.springframework.data.jpa.repository.JpaRepository;
